% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/featuresExtractionTS.r
\docType{data}
\name{featuresExtractionTS}
\alias{featuresExtractionTS}
\title{Features Extraction from Time Series}
\usage{
featuresExtractionTS(fileTS = NULL, nameColumnValue = NULL, 
subInterval = FALSE, numberSubIntervals = 2)
}
\arguments{
\item{fileTS}{Dataframe. A time series file.}

\item{nameColumnValue}{Character. Name of a column with value, such as, Vegetation Index EVI or NDVI.}

\item{subInterval}{Logical. If TRUE, the data.frame will be divide in subintervals. Given a data.frame with 23 rows and 3 subintervals, the first and second subinterval will have 8 rows each, and the last only 7 rows. If FALSE (the default) nothing subinterval will be created.}

\item{numberSubIntervals}{Integer. Number of subintervals to feature extraction.}
}
\value{
Dataframe with statistical features
}
\description{
Performed a feature extraction from time series data. And return statistical summary for time series entire or by intervals of the time series
}
\examples{
\dontrun{
# Open a data example
library(featuresTS)
data("dataTS")
df <- dataTS

# Apply the filterTS function on df data frame 
dataFiltered <- filterTS(fileTS = df, nameColumnValue = "value", outlier = TRUE, value= -0.300)

# Apply splitTS fcuntion to divide time series for year 
splitTS(dataFiltered,2002,2005,"date",typeInterval = "annual")
 
# Get features from time series divided in annual values without subintervals 
example1 <- featuresExtractionTS(fileTS = ts.annual_2002, nameColumnValue = 
"filtered.value", subInterval = FALSE)

# Get features from time series divided in annual values with subintervals 
example2 <- featuresExtractionTS(fileTS = ts.annual_2002, nameColumnValue = 
"filtered.value", subInterval = TRUE, numberSubIntervals = 3)

# Show data frames example1 and example2 utils::View(example1) utils::View(example2) 

}

}
\author{
Adeline M. Maciel
}
\keyword{datasets}

