% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/filterTS.r
\docType{data}
\name{filterTS}
\alias{filterTS}
\title{Filter Time Series using Savitzky-Golay filter}
\usage{
filterTS(fileTS = NULL, nameColumnValue = NULL, outlier = FALSE, value = -0.3000)
}
\arguments{
\item{fileTS}{Dataframe. A file time series}

\item{nameColumnValue}{Character. A name of column with value to filter}

\item{outlier}{Logical. If FALSE (the default), there none outlier for to make interpolation. If TRUE, there some gaps with values to make interpolation.}

\item{value}{Numeric. A number, double or integer, if outlier is TRUE. The default is -0.3000}
}
\value{
Dataset with new column filtered
}
\description{
Return a new data.frame with values of time series filtered.
}
\examples{
\dontrun{
# Open a data example
library(featuresTS)
data("dataTS")
df <- dataTS
 
# The time series data contains 4 columns and 207 rows like
# longitude  latitude       date   value
#  -57.0474 -11.36449 2004-06-25  0.3125
#  -57.0474 -11.36449 2004-07-11  0.3235
#  -57.0474 -11.36449 2004-07-27 -0.3000
#  -57.0474 -11.36449 2004-08-12  0.3648

# Apply filterTS function 
dataFiltered <- filterTS(fileTS = df, nameColumnValue = "value", outlier = TRUE, value= -0.3000)

# Show new data.frame with values filtered
head(dataFiltered)

}

}
\author{
Adeline M. Maciel
}
\seealso{
Savitzky-Golay filter details in \code{\link{signal}} package
}
\keyword{datasets}

