% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/focalFeaturesTS.R
\docType{data}
\name{focalFeaturesTS}
\alias{focalFeaturesTS}
\title{Focal Features Extraction from Time Series}
\usage{
focalFeaturesTS(fileTS = NULL, valueToleranceRaster = 0.00000)
}
\arguments{
\item{fileTS}{A file time series alredy with features extracted and divided for year ou subinterval.}

\item{valueToleranceRaster}{A number of tolerance case raster of the time series contains coordinate intervals not constant. Because this function trasforme time series for raster data. Default is 0.00000.}
}
\value{
new dataset of data.frame with features of the focal neighborhood statistical
}
\description{
Performed a focal feature extraction from time series data. And return focal statistical summary for time series divided for year. Calculates for each input time series, cell  location, a statistic of the values within a specified neighborhood should be 8 around it (Queen's case).
}
\examples{
\dontrun{
# open a data example
library(featuresTS)
data("dataFeaturesTS")
df <- dataFeaturesTS

# features extraction for entire data.frame without subIntervals
focalFeaturesTS(fileTS = df, valueToleranceRaster = 0.000891266)
# result in data.frame with focal features

}

}
\author{
Adeline M. Maciel
}
\keyword{datasets}

