% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/featuresExtractionTS.r
\docType{data}
\name{featuresExtractionTS}
\alias{featuresExtractionTS}
\title{Features Extraction from Time Series}
\usage{
featuresExtractionTS(fileTS = NULL, nameColumnValue = NULL, 
subInterval = FALSE, numberSubIntervals = 2)
}
\arguments{
\item{fileTS}{A file time series.}

\item{nameColumnValue}{A name of column with value, such as, vegetation indice}

\item{subInterval}{If TRUE, the data.frame will be divide in subintervals, for example, given a data.frame with 23 rows and 3 subintervals, the first and second subinterval will have 8 rows each, and the last only 7 rows. Default is FALSE.}

\item{numberSubIntervals}{A number of subintervals.}
}
\value{
new dataset of data.frame with statistical features
}
\description{
Performed a feature extraction from time series data. And return statistical summary for time series entire or by intervals of the time series
}
\examples{
\dontrun{
# open a data example
library(featuresTS)
data("dataTS")
df <- dataTS

# features extraction for entire data.frame without subIntervals
featuresExtractionTS(fileTS = df, nameColumnValue = "value", 
subInterval = FALSE)
# result in data.frame with statistical features

# features extraction for entire data.frame and  2 subIntervals
featuresExtractionTS(fileTS = df, nameColumnValue = "value", 
subInterval = TRUE, numberSubIntervals = 2)

}

}
\author{
Adeline M. Maciel
}
\keyword{datasets}

