% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/filterTS.r
\docType{data}
\name{filterTS}
\alias{filterTS}
\title{Filter Time Series using Savitzky-Golay filter}
\usage{
filterTS(fileTS = NULL, nameColumnValue = NULL, outlier = FALSE, value = -0.3000)
}
\arguments{
\item{fileTS}{A file time series}

\item{nameColumnValue}{A name of column with value to filter}

\item{outlier}{TRUE/FALSE if have any outlier for to make interpolation}

\item{value}{A single integer if outlier is TRUE}
}
\value{
a new data.frame, called data.filtered, with new column of the value filtered
}
\description{
Return a new data.frame with values of time series filtered.
}
\examples{
\dontrun{
# open a data example
library(featuresTS)
data("dataTS")
df <- dataTS
 
# the time series data contains 4 columns and 207 rows like
# longitude  latitude       date   value
#  -57.0474 -11.36449 2004-06-25  0.3125
#  -57.0474 -11.36449 2004-07-11  0.3235
#  -57.0474 -11.36449 2004-07-27 -0.3000
#  -57.0474 -11.36449 2004-08-12  0.3648

# apply function filterTS
filterTS(fileTS = df, nameColumnValue = "value", outlier = TRUE, value= -0.300)

# show new data.frame with values filtered
head(data.filtered)
}


}
\author{
Adeline M. Maciel
}
\seealso{
Savitzky-Golay filter details in \code{\link{signal}} package
}
\keyword{datasets}

