% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/filterTS.r
\name{filterTS}
\alias{filterTS}
\title{Filter Time Series using Savitzky–Golay filter}
\usage{
filterTS ( fileTS = data, nameColumnValue = "V1", outlier = if(TRUE)
define a number to next parameter value, value = integer )
}
\arguments{
\item{fileTS}{file time series}

\item{nameColumnValue}{name of column with value to filter}

\item{outlier}{TRUE/FALSE if have any outlier for to make interpolation}

\item{value}{a single integer if outlier is TRUE}
}
\value{
data.filtered new data.frame with new column with value filtered
}
\description{
Return a new data.frame with values of time series filtered.
}
\examples{

# open a data example
df <- read.csv("data/dataTimeSeries.csv")

# the time series data contains 4 columns and 207 rows like
# longitude  latitude       date   value
#  -57.0474 -11.36449 2004-06-25  0.3125
#  -57.0474 -11.36449 2004-07-11  0.3235
#  -57.0474 -11.36449 2004-07-27 -0.3000
#  -57.0474 -11.36449 2004-08-12  0.3648

# apply function filterTS
library(featuresTS)
filterTS(fileTS = df, nameColumnValue = "value", outlier = TRUE, value= -0.300)

# show new data.frame with values filtered
head(data.filtered)




}
\seealso{
Savitzky–Golay filter details in \code{\link{signal}} package
}
\keyword{Savitzky–Golay}
\keyword{extraction,}
\keyword{features}
\keyword{filter}
\keyword{series,}
\keyword{time}

