% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/focalFeaturesTS.R
\docType{data}
\name{focalFeaturesTS}
\alias{focalFeaturesTS}
\title{Focal Features Extraction from Time Series}
\usage{
focalFeaturesTS(fileTS = NULL, valueToleranceRaster = 0.00000)
}
\arguments{
\item{fileTS}{Dataframe. A file time series alredy with features extracted and divided for year ou subinterval.}

\item{valueToleranceRaster}{Integer. A number of tolerance case raster of the time series contains coordinate intervals not constant. Because this function trasforme a time series in raster object to calcule neighborhood. Default is 0.00000.}
}
\value{
Dataset with features of focal neighborhood statistical
}
\description{
Performed a focal feature extraction from time series data. And return focal statistical summary for time series divided for year. Calculates for each input time series, cell  location, a statistic of the values within a specified neighborhood should be 8 around it (Queen's case).
}
\note{
This function requires that the igraph package is available.
}
\examples{
\dontrun{
# Open a data example
library(featuresTS)
data("dataFeaturesTS")
df <- dataFeaturesTS

# Load igraph package
library(igraph)

# Focal neighborhood features extraction for dataset of time series 
dfTSwithFocalFeatures <- focalFeaturesTS(fileTS = df, valueToleranceRaster = 0.000891266)

head(dfTSwithFocalFeatures)

}

}
\author{
Adeline M. Maciel
}
\keyword{datasets}

